Move ArrayFetcher to tag_fetcher and expose it
